refactor(views): create base class for preview menu properties
